# openapi.yaml
openapi: 3.0.3
info:
  title: DRP · Domestic Resource Planning API
  description: >
    API REST del core de DRP (Domestic Resource Planning).


    Este documento es la **fuente de verdad** de la API, no una ilustración
    de ella (ver ADR-007): se edita antes que el código que lo implementa, el
    frontend genera de él sus tipos y su cliente, y el backend se verifica
    contra él con las pruebas de contrato. Ver readme.md del proyecto para el
    contexto completo — modelo de dominio, event bus y roadmap.


    Dos convenciones transversales. Los **errores de forma** responden 400 con
    código `VALIDATION_ERROR`; los de **regla de negocio**, 409 con el código
    concreto. Y todas las colecciones se devuelven **paginadas**, con la misma
    envoltura `{ items, page, size, total }`.
  version: "0.3.0-fase2"
  license:
    name: Ver LICENSE del repositorio
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local, levantado por compose.yaml
  - url: https://api.drp.example/api/v1
    description: Entorno ilustrativo (el real se fija al elegir el VPS, al cerrar la Fase 1)
...
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Listar las categorías del hogar
      parameters:
        - name: includeRetired
          in: query
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Listado de categorías vigentes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    post:
      operationId: createCategory
      tags: [categories]
      summary: Crear una categoría
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "201":
          description: Categoría creada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "409":
          description: Ya existe una categoría vigente con ese name (CATEGORY_DUPLICATE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Renombrar o anotar una categoría
      description: >
        El nombre sigue siendo único entre las vigentes del hogar, comparado
        normalizado. Renombrar no toca a los assets ni a los artículos que la
        usan: la referencian por identificador, no por texto.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "200":
          description: Categoría actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Ya hay otra categoría vigente con ese nombre (CATEGORY_DUPLICATE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    delete:
      operationId: retireCategory
      tags: [categories]
      summary: Retirar una categoría
      description: >
        Retirada lógica. Deja de ofrecerse al clasificar, pero los assets y
        artículos que ya la tenían la conservan, así que la fila permanece.
      responses:
        "204":
          description: Categoría retirada
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: Listar y buscar artículos del catálogo del hogar
      description: >
        Alimenta el autocompletado del alta de consumibles: se busca el artículo
        antes de dar entrada, para no volver a introducir su definición.
      parameters:
        - name: q
          in: query
          description: Búsqueda por name (normalizado, sin distinguir mayúsculas ni acentos)
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: barcode
          in: query
          schema: { type: string }
        - name: includeRetired
          in: query
          description: Por defecto false -- los artículos retirados no se ofrecen en el alta
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Listado de artículos vigentes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticlePage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    post:
      operationId: createArticle
      tags: [articles]
      summary: Crear un artículo en el catálogo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleInput" }
      responses:
        "201":
          description: Artículo creado. Publica el evento ArticleCreated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "409":
          description: >
            Ya existe un artículo con ese name normalizado o ese código de
            barras en el hogar (ARTICLE_DUPLICATE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /articles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getArticle
      tags: [articles]
      summary: Detalle de un artículo
      responses:
        "200":
          description: Artículo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Corregir la ficha de un artículo
      description: >
        Cambia la definición, no las existencias. La unidad es el único campo
        que deja de poder cambiarse en cuanto el artículo tiene existencias
        vivas: alterarla reinterpretaría en silencio toda la cantidad ya
        contada, que se lleva en esa unidad.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticlePatch" }
      responses:
        "200":
          description: Artículo actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Ya hay otro artículo vigente con ese nombre o ese código de barras
            (ARTICLE_DUPLICATE), o se intenta cambiar la unidad de uno que ya
            tiene existencias (ARTICLE_UNIT_IMMUTABLE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    delete:
      operationId: retireArticle
      tags: [articles]
      summary: Retirar un artículo del catálogo
      description: >
        Retirada lógica -- marca retiredAt. La fila permanece porque las
        existencias dadas de baja siguen apuntando a ella; borrarla rompería
        la clave ajena y el historial. Un artículo retirado no se ofrece en el
        alta ni admite nuevas entradas.
      responses:
        "204":
          description: Artículo retirado
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Al artículo aún le quedan existencias vivas
            (ARTICLE_HAS_EXISTENCES)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de usuarios del hogar. Claims -- sub (identityId), memberId
        (pertenencia en el hogar), householdId, role y sid. La distincion entre
        sub y memberId importa: todo lo que el dominio llama "usuario"
        --propietario de un asset, prestador, receptor, createdBy-- apunta a la
        pertenencia, mientras que los refresh tokens cuelgan de la identidad.


        sid identifica cual de las sesiones de esa persona es esta: es el
        identificador del refresh token con el que se emitio. Existe porque
        POST /auth/password revoca las DEMAS sesiones y conserva la que esta en
        uso, y para conservarla hay que saber cual es. La alternativa era pedir
        el refresh token en el cuerpo de esa operacion, que obligaria al cliente
        a mandar una credencial de vida larga donde no hace falta. No es un
        secreto: identifica una fila cuyo valor real esta hasheado.
    loanToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Credencial de un prestador o receptor **externo**, sin cuenta en DRP.
        Es un JWT firmado **sin sub de usuario**, con claims loanId y role
        (LENDER | BORROWER), que viaja como enlace por correo al iniciar el
        prestamo.


        Su alcance son exactamente dos operaciones -- GET /loans/{id} de ESE
        prestamo y POST /loans/{id}/return -- y ninguna otra cosa del hogar.
        Ni el listado, ni otro prestamo, ni los assets.


        Vale a la vez la firma y la fila: su hash se guarda en
        loan_access_tokens, y verificarlo exige firma valida **y** fila viva.
        Las dos capas hacen cosas distintas y por eso estan las dos -- la firma
        lo hace infalsificable y deja rechazar lo caducado sin tocar la base de
        datos; la fila lo hace **revocable**, que un JWT por si solo no puede
...
  parameters:
...
    Page:
      name: page
      in: query
      description: Pagina solicitada, empezando en 0
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: Elementos por pagina
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Falta el token, ha caducado o no es válido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: >
        El cuerpo o los parámetros no cumplen el contrato (VALIDATION_ERROR).
        Es el error de forma; los de negocio van en 409. `details` lleva un
        campo por atributo rechazado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: >
        Error no previsto (INTERNAL_ERROR). El cuerpo nunca lleva detalle
        interno: la traza va al log, no a la respuesta.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      description: >
        Forma única de todo error de la API. `code` es lo que el cliente trata;
        `message` es texto para diagnóstico, no para mostrar tal cual.
      properties:
        code:
          type: string
          example: ASSET_LOCATION_CONFLICT
          description: >
            Enumerado a propósito: el cliente decide qué hacer por el código, y
            un código que no esté aquí es un fallo del contrato, no un caso que
            el frontend deba adivinar.
          enum:
            - ALREADY_MEMBER
            - ARTICLE_DUPLICATE
            - ARTICLE_HAS_EXISTENCES
            - ARTICLE_UNIT_IMMUTABLE
            - ASSET_HAS_ACTIVE_LOAN
            - ASSET_HAS_CHILDREN
            - ASSET_LOCATION_CONFLICT
            - ASSET_QUANTITY_NEGATIVE
            - ASSET_QUANTITY_NOT_APPLICABLE
            - CATEGORY_DUPLICATE
            - CURRENT_PASSWORD_INVALID
            - DOCUMENT_CONTENT_INVALID
            - DOCUMENT_TARGET_INVALID
            - EMAIL_NOT_VERIFIED
            - EXISTENCE_ALREADY_IN_LOCATION
            - FILE_ALREADY_ATTACHED
            - FILE_IN_USE
            - FILE_TOO_LARGE
            - FILE_TYPE_NOT_ALLOWED
            - IDENTITY_ALREADY_MEMBER
            - INTAKE_QUANTITY_NOT_POSITIVE
            - INTERNAL_ERROR
            - INVITATION_ALREADY_PENDING
            - INVITATION_TOKEN_INVALID
            - LOAN_ALREADY_RETURNED
            - LOAN_ASSET_ALREADY_LENT
            - LOAN_ASSET_NOT_DURABLE
            - LOCATION_CYCLE
            - LOCATION_DUPLICATE
            - LOCATION_HAS_ASSETS
            - LOCATION_HAS_CHILDREN
            - MERGE_ARTICLE_MISMATCH
            - MERGE_ASSET_DEACTIVATED
            - MERGE_NOT_CONSUMABLE
            - MERGE_SAME_ASSET
            - MODULE_INACTIVE
            - RATE_LIMITED
            - RESET_TOKEN_INVALID
            - STORAGE_QUOTA_EXCEEDED
            - USER_LAST_ADMIN
            - VALIDATION_ERROR
            - VERIFICATION_TOKEN_INVALID
        message: { type: string }
        details:
          type: object
          description: Contexto adicional del error; su forma depende del código
...
    PageMeta:
      type: object
      required: [page, size, total]
      description: >
        Metadatos de paginación, comunes a todas las colecciones. Son las mismas
        en las doce sin excepción por tamaño esperado: una sola forma que
        aprender en el cliente, y ninguna migración el día que una lista que se
        creía pequeña deje de serlo.
      properties:
        page: { type: integer, minimum: 0, example: 0, description: "Página devuelta, empezando en 0" }
        size: { type: integer, minimum: 1, example: 50, description: Tamaño de página aplicado }
        total: { type: integer, format: int64, example: 137, description: Total de elementos que cumplen el filtro }
...
    Authorship:
      type: object
      description: >
        Autoria de la fila, solo lectura. Se toma del token de quien hace la
        peticion y se ignora si el cliente la envia. Nulo significa que el
        cambio lo hizo el sistema y no una persona -- el caso del proceso
        diario de prestamos vencidos.
      properties:
        createdBy: { type: string, format: uuid, nullable: true }
        updatedBy: { type: string, format: uuid, nullable: true }
...
    MeasurementUnit:
      type: string
      description: >
        Unidad de medida de un artículo del catálogo. La fija el artículo, no la
        existencia: todas las existencias de un mismo artículo se llevan en ella.
      enum: [UNIT, GRAM, KILOGRAM, MILLILITER, LITER, METER, PACK]
    CategoryInput:
      type: object
      description: >
        Clasificación funcional del hogar. Cada hogar arranca con un juego
        sembrado y lo edita a su gusto.
      required: [name]
      properties:
        name:
          type: string
          description: Único entre las categorías vigentes del hogar, comparado normalizado
        notes: { type: string }
    Category:
      allOf:
        - $ref: "#/components/schemas/CategoryInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            retiredAt:
              type: string
              format: date-time
              nullable: true
              description: Informado si la categoría está retirada
        - $ref: "#/components/schemas/Authorship"
...
    ArticleInput:
      type: object
      description: >
        Definición reutilizable de un artículo del hogar. No es un asset -- no
        tiene ubicación ni cantidad.
      required: [name, categoryId, unit]
      properties:
        name:
          type: string
          description: Único en el hogar, comparado normalizado (sin mayúsculas ni acentos)
        categoryId:
          type: string
          format: uuid
          description: Categoría vigente del catálogo del hogar
        unit: { $ref: "#/components/schemas/MeasurementUnit" }
        brand: { type: string }
        model:
          type: string
          description: Marca y modelo juntos son la identidad real de un duradero
        barcode:
          type: string
          description: Opcional; único en el hogar si se informa
        packSize:
          type: number
          exclusiveMinimum: true
          minimum: 0
          description: >
            Contenido de un envase, en la unidad del artículo. Permite dar
            entrada a "dos paquetes" sin multiplicar a mano.
        photoUrl:
          type: string
          format: uri
          description: >
            En escritura, enlace externo a la imagen. En lectura, la URL con la
            que se pinta: ese mismo enlace, o la URL FIRMADA de vida corta del
            fichero subido, que caduca en unos 15 minutos y no debe guardarse
            (ver readme.md 5.8.4).
        photoFileId:
          type: string
          format: uuid
          description: >
            Foto subida al servidor. Excluyente con photoUrl -- se admite una
            de las dos o ninguna, porque no tener foto es lo normal.
        notes: { type: string }
      not:
        required: [photoUrl, photoFileId]
    ArticlePatch:
      type: object
      description: >
        Todos los campos son opcionales; solo se aplica lo que llegue. unit no
        se acepta si el artículo ya tiene existencias vivas -- responde
        ARTICLE_UNIT_IMMUTABLE--, porque la cantidad ya contada se lleva en ella.
      properties:
        name: { type: string }
        categoryId: { type: string, format: uuid }
        unit: { $ref: "#/components/schemas/MeasurementUnit" }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        packSize:
          type: number
          nullable: true
          exclusiveMinimum: true
          minimum: 0
        photoUrl: { type: string, format: uri, nullable: true }
        photoFileId:
          type: string
          format: uuid
          nullable: true
          description: Nulo desadjunta la foto sin borrar el fichero
        notes: { type: string, nullable: true }
      not:
        required: [photoUrl, photoFileId]
    Article:
      allOf:
        - $ref: "#/components/schemas/ArticleInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            category:
              type: string
              description: Nombre de la categoría, resuelto para lectura
            photoThumbnailUrl:
              type: string
              format: uri
              nullable: true
              description: Miniatura firmada, para listados. Solo si la foto es un fichero subido
            createdAt: { type: string, format: date-time }
            retiredAt:
              type: string
              format: date-time
              nullable: true
              description: Informado si el artículo está retirado del catálogo
        - $ref: "#/components/schemas/Authorship"
...
    ArticlePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Article" }
...
    CategoryPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Category" }
...